+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Remove `database.backup()` in favour of searching for a solution disconnected from the bot.
 
+### Features
+
+- Improve database caching.
+
 ## Version 0.7.8
 
 **[Tag Comparison](https://github.com/JamCoreDiscord/Pinguino/compare/v0.7.7...v0.7.8)**